Added: ignoring cases.

+++ autotest/run_test.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 PYTHON=python2
-SYNCDIR="$PYTHON run_syncdir.py -c../syncdir.py -B_backup -T_trash"
+SYNCDIR="$PYTHON run_syncdir.py -c../syncdir.py -i -B_backup -T_trash"
 TESTBASE=testdir
 
 # create test directories
@@ -25,6 +25,7 @@ cmp $D1/ttt/t $D2/ttt/t || exit 103
 echo bar > $D2/bar
 $SYNCDIR $D1 $D2
 cmp $D1/bar $D2/bar || exit 104
+sleep 1
 
 # ignoring files
 echo "*** ignoring files"
@@ -35,10 +36,10 @@ $SYNCDIR -Iyyy $D1 $D2
 [ ! -f $D2/.xxx ] || exit 301
 [ ! -f $D2/a.yyy ] || exit 302
 cmp $D1/a.zzz $D2/a.zzz || exit 303
+sleep 1
 
 # updating files
 echo "*** updating files"
-sleep 1
 cp $D1/foo $TESTDIR/foo
 echo foooo > $D1/foo
 cp $D1/bar $TESTDIR/bar
@@ -48,10 +49,23 @@ cmp $D1/bar $TESTDIR/bar || exit 201
 cmp $D1/bar $D2/bar || exit 202
 cmp $D1/foo $D2/foo || exit 203
 cmp $TESTDIR/foo $D2/_backup/foo.backup.* || exit 304
+sleep 1
+
+# ignoring cases
+echo "*** ignoring cases"
+mv $D1/foo $D1/FOO
+mv $D2/bar $D2/Bar
+$SYNCDIR $D1 $D2
+[ ! -f $D2/FOO ] || exit 401
+[ ! -f $D1/Bar ] || exit 402
+cmp $D1/FOO $D2/foo || exit 403
+cmp $D1/bar $D2/Bar || exit 404
+mv $D1/FOO $D1/foo
+mv $D2/Bar $D2/bar
+sleep 1
 
 # trashing files
 echo "*** trashing files"
-sleep 1
 mkdir $D1/_trash
 mkdir $D2/_trash
 mv $D1/foo $D1/_trash/foo
@@ -61,6 +75,8 @@ $SYNCDIR $D1 $D2
 [ ! -f $D2/foo ] || exit 501
 [ ! -f $D2/ttt ] || exit 502
 [ ! -f $D1/bar ] || exit 503
+sleep 1
 
 # cleanup
+#echo "*** cleanup"
 #rm -r $TESTDIR
